Fix broken iSCSI shared storage support.

+++ dcmgr/lib/dcmgr/rpc/sta_handler.rb
@@ -10,58 +10,68 @@ module Dcmgr
       include Dcmgr::Helpers::CliHelper
       include Dcmgr::Helpers::ByteUnit
 
-      def select_backing_store
-        backing_store = Dcmgr.conf.backing_store
-        @backing_store = Dcmgr::Drivers::BackingStore.select_backing_store(backing_store)
+      def backing_store
+        @backing_store = Dcmgr::Drivers::BackingStore.driver_class(Dcmgr.conf.backing_store_driver).new
       end
 
-      def select_iscsi_target
-        iscsi_target = Dcmgr.conf.iscsi_target
-        @iscsi_target = Dcmgr::Drivers::IscsiTarget.select_iscsi_target(iscsi_target, @node)
+      def iscsi_target
+        @iscsi_target = Dcmgr::Drivers::IscsiTarget.driver_class(Dcmgr.conf.iscsi_target_driver).new
       end
 
       # Setup volume file from snapshot storage and register to
       # sotrage target.
       def setup_and_export_volume
-        select_backing_store
         @sta_ctx = StaContext.new(self)
 
-        rpc.request('sta-collector', 'update_volume', @volume_id, {:state=>:creating, :export_path=>@volume[:uuid]})
+        rpc.request('sta-collector', 'update_volume', @volume_id, {:state=>:creating})
 
         if @volume[:backup_object_id]
-          begin
-            snap_tmp_path = File.expand_path("#{@volume[:uuid]}.tmp", Dcmgr.conf.tmp_dir)
 
-            @backup_object = @snapshot = rpc.request('sta-collector', 'get_backup_object', @volume[:backup_object_id])
-            raise "Invalid backup_object state: #{@backup_object[:state]}" unless @backup_object[:state].to_s == 'available'
+          @backup_object = rpc.request('sta-collector', 'get_backup_object', @volume[:backup_object_id])
+          raise "Invalid backup_object state: #{@backup_object[:state]}" unless @backup_object[:state].to_s == 'available'
+          
+          if backing_store.local_backup_object?(@backup_object)
+            # the backup data exists on the same storage and also
+            # is known how to convert to volume by the backing
+            # store driver. e.g. filesystem level snapshot.
+
+            logger.info("Creating new volume #{@volume_id} from #{@backup_object[:uuid]} (#{convert_byte(@volume[:size], MB)} MB)")
+            backing_store.create_volume(@sta_ctx, @backup_object[:object_key])
+            
+          else
+            # download backup data from backup storage. then create
+            # volume from the snapshot.
 
             begin
-              # download backup object to the tmporary place.
-              snapshot_storage = Drivers::BackupStorage.snapshot_storage(@backup_object[:backup_storage])
-              logger.info("Downloading to #{@backup_object[:uuid]}: #{snap_tmp_path}")
-              snapshot_storage.download(@backup_object, snap_tmp_path)
-              logger.info("Finished downloading #{@backup_object[:uuid]}: #{snap_tmp_path}")
-            rescue => e
-              logger.error(e)
-              raise "snapshot not downloaded"
+              snap_tmp_path = File.expand_path("#{@volume[:uuid]}.tmp", Dcmgr.conf.tmp_dir)
+              
+              begin
+                # download backup object to the tmporary place.
+                backup_storage = Drivers::BackupStorage.snapshot_storage(@backup_object[:backup_storage])
+                logger.info("Downloading to #{@backup_object[:uuid]}: #{snap_tmp_path}")
+                backup_storage.download(@backup_object, snap_tmp_path)
+                logger.info("Finished downloading #{@backup_object[:uuid]}: #{snap_tmp_path}")
+              rescue => e
+                logger.error(e)
+                raise "Failed to download backup object: #{@backup_object[:uuid]}"
+              end
+              logger.info("Creating new volume #{@volume_id} from #{@backup_object[:uuid]} (#{convert_byte(@volume[:size], MB)} MB)")
+              
+              backing_store.create_volume(@sta_ctx, snap_tmp_path)
+            ensure
+              File.unlink(snap_tmp_path) rescue nil
             end
-            logger.info("Creating new volume #{@volume_id} from #{@backup_object[:uuid]} (#{convert_byte(@volume[:size], MB)} MB)")
-
-            @backing_store.create_volume(@sta_ctx, snap_tmp_path)
-          ensure
-            File.unlink(snap_tmp_path) rescue nil
           end
-
+          
         else
-          logger.info("Creating new empty volume #{@volume_id} (#{convert_byte(@volume[:size], MB)} MB)")
-          @backing_store.create_volume(@sta_ctx, nil)
+          logger.info("Creating new blank volume #{@volume_id} (#{convert_byte(@volume[:size], MB)} MB)")
+          backing_store.create_volume(@sta_ctx, nil)
         end
         logger.info("Finished creating new volume #{@volume_id}.")
 
         logger.info("Registering to iscsi target: #{@volume_id}")
-        select_iscsi_target
-        opt = @iscsi_target.create(@sta_ctx)
-        rpc.request('sta-collector', 'update_volume', @volume_id, {:state=>:available, :transport_information=>opt})
+        opt = iscsi_target.create(@sta_ctx)
+        rpc.request('sta-collector', 'update_volume', @volume_id, {:state=>:available, :volume_device=>opt})
         logger.info("Finished registering iscsi target: #{@volume_id}")
       end
 
@@ -88,7 +98,7 @@ module Dcmgr
         setup_and_export_volume
 
         @instance = rpc.request('hva-collector', 'get_instance', @instance_id)
-        jobreq.submit("hva-handle.#{@instance[:host_node][:node_id]}", 'run_vol_store', @instance_id, @volume_id)
+        jobreq.submit("hva-handle.#{@instance[:host_node][:node_id]}", 'run_vol_store', @instance_id)
       }, proc {
         # TODO: need to clear generated temp files or remote files in remote snapshot repository.
         rpc.request('sta-collector', 'update_volume', @volume_id, {:state=>:deleted, :deleted_at=>Time.now.utc})
@@ -108,33 +118,57 @@ module Dcmgr
           logger.warn("#{@volume_id}: Unexpected volume state but try destroy resource: #{@volume[:state]}")
         end
 
-        # deregisterd iscsi target
-        select_iscsi_target
+        rpc.request('sta-collector', 'update_volume', @volume_id, {:state=>:deleting})
+
+        # deregister iscsi target
         begin
-          @iscsi_target.delete(StaContext.new(self))
+          iscsi_target.delete(StaContext.new(self))
         rescue => e
           logger.error("#{@volume_id}: Failed to delete ISCSI target entry.")
+          logger.error(e)
           errcount += 1
         end
 
-        rpc.request('sta-collector', 'update_volume', @volume_id, {:state=>:deleting})
-
         # delete volume
-        select_backing_store
         begin
-          @backing_store.delete_volume(StaContext.new(self))
+          backing_store.delete_volume(StaContext.new(self))
         rescue => e
-          logger.error("#{@volume_id}: Failed to delete volume: #{@volume[:storage_node][:export_path]}/#{@volume[:uuid]}")
+          logger.error("#{@volume_id}: Failed to delete volume.")
+          logger.error(e)
           errcount += 1
         end
 
         rpc.request('sta-collector', 'update_volume', @volume_id, {:state=>:deleted, :deleted_at=>Time.now.utc})
         if errcount > 0
-          logger.info("#{@volume_id}: Encountered one or more errors during deleting.")
+          raise "#{@volume_id}: Encountered one or more errors during deleting."
         else
           logger.info("#{@volume_id}: Deleted volume successfully.")
         end
       end
+
+      job :backup_volume, proc {
+        @volume_id = request.args[0]
+        @backup_object_id = request.args[1]
+        @volume = rpc.request('sta-collector', 'get_volume', @volume_id)
+        @backup_object = rpc.request('sta-collector', 'get_backup_object', @backup_object_id) unless @backup_object_id.nil?
+        @sta_ctx = StaContext.new(self)
+
+        # backup volume
+        if backing_store.kind_of?(Dcmgr::Drivers::BackingStore::ProvideBackupVolume)
+          backing_store.backup_volume(@sta_ctx)
+          new_object_key = backing_store.backup_object_key_created(@sta_ctx)
+        elsif backing_store.kind_of?(Dcmgr::Drivers::BackingStore::ProvidePointInTimeSnapshot)
+          # take one generation snapshot -> copy data -> delete snapshot.
+          new_object_key = nil
+          raise NotImplementedError
+        else
+          raise "None of backup operation types are supported by #{backing_store.class}."
+        end
+
+        rpc.request('sta-collector', 'update_backup_object', @backup_object_id,
+                    {:state=>:available, :object_key=>new_object_key})
+        logger.info("created new backup: #{@backup_object_id}")
+      }
 
       job :create_snapshot, proc {
         @volume_id = request.args[0]
@@ -148,24 +182,22 @@ module Dcmgr
 
         begin
           snapshot_storage = Dcmgr::Drivers::BackupStorage.snapshot_storage(@backup_object[:backup_storage])
-          select_backing_store
 
           logger.info("Taking new snapshot for #{@volume_id}")
-          @backing_store.create_snapshot(@sta_ctx)
+          backing_store.create_snapshot(@sta_ctx)
           logger.info("Finish to create snapshot for #{@volume_id}")
           logger.info("Uploading #{@backup_object_id} to #{@backup_object[:backup_storage][:base_uri]}")
-          snapshot_storage.upload(@backing_store.snapshot_path(@sta_ctx), @backup_object)
+          snapshot_storage.upload(backing_store.snapshot_path_created(@sta_ctx), @backup_object)
           logger.info("Finish to upload #{@backup_object_id}")
         rescue => e
           logger.error(e)
           raise "snapshot has not be uploaded"
         ensure
-          @backing_store.delete_snapshot(@sta_ctx)
+          backing_store.delete_snapshot(@sta_ctx)
         end
 
-        rpc.request('sta-collector', 'update_backup_object', @backup_object_id, {:state=>:available}) do |req|
-          req.oneshot = true
-        end
+        rpc.request('sta-collector', 'update_backup_object', @backup_object_id,
+                    {:state=>:available, :path=>backing_store.snapshot_path_created(@sta_ctx)})
         logger.info("created new backup: #{@backup_object_id}")
       }, proc {
         # TODO: need to clear generated temp files or remote files in remote snapshot repository.
